Remove some more duplications
