+++ src/main/java/io/antmedia/rest/BroadcastRestServiceV2.java
@@ -470,32 +470,9 @@ public class BroadcastRestServiceV2 extends RestServiceBase{
 	@GET
 	@Path("/{id}/broadcast-statistics")
 	@Produces(MediaType.APPLICATION_JSON)
+	@Override
 	public BroadcastStatistics getBroadcastStatistics(@ApiParam(value = "the id of the stream", required = true) @PathParam("id") String id) {
-
-		int totalRTMPViewer = -1;
-		int totalWebRTCViewer = -1;
-		int totalHLSViewer = -1;
-		if (id != null) 
-		{
-			IBroadcastScope broadcastScope = getScope().getBroadcastScope(id);
-
-			if (broadcastScope != null)	{
-				totalRTMPViewer = broadcastScope.getConsumers().size();
-			}
-
-			Broadcast broadcast = getDataStore().get(id);
-			if (broadcast != null) {
-				totalHLSViewer = broadcast.getHlsViewerCount();
-			}
-
-			IWebRTCAdaptor webRTCAdaptor = getWebRTCAdaptor();
-
-			if (webRTCAdaptor != null) {
-				totalWebRTCViewer = webRTCAdaptor.getNumberOfViewers(id);
-			}
-		}
-
-		return new BroadcastStatistics(totalRTMPViewer, totalHLSViewer, totalWebRTCViewer);
+		return super.getBroadcastStatistics(id);
 	}
 
 	@ApiOperation(value = "Get WebRTC Client Statistics such as : Audio bitrate, Video bitrate, Target bitrate, Video Sent Period etc.", notes = "", responseContainer = "List",response = WebRTCClientStats.class)
@@ -527,86 +504,9 @@ public class BroadcastRestServiceV2 extends RestServiceBase{
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Path("/social-networks/{serviceName}")
 	@Produces(MediaType.APPLICATION_JSON)
+	@Override
 	public Object getDeviceAuthParameters(@ApiParam(value = "Name of the service, like Facebook, Youtube, Periscope", required = true) @PathParam("serviceName") String serviceName) {
-		String message = null;
-		boolean missingClientIdAndSecret = false;
-
-		int errorId = -1;
-		VideoServiceEndpoint videoServiceEndpoint = null;
-		if (serviceName.equals(AntMediaApplicationAdapter.FACEBOOK)) 
-		{
-			String clientId = getAppSettings().getFacebookClientId();
-			String clientSecret = getAppSettings().getFacebookClientSecret();
-
-			videoServiceEndpoint = getApplication().getEndpointService(AntMediaApplicationAdapter.FACEBOOK_ENDPOINT_CLASS, null, clientId, clientSecret);
-
-			if (isClientIdMissing(videoServiceEndpoint, clientId, clientSecret)) 
-			{
-				missingClientIdAndSecret = true;
-			}
-
-		}
-		else if (serviceName.equals(AntMediaApplicationAdapter.YOUTUBE)) 
-		{
-
-			String clientId = getAppSettings().getYoutubeClientId();
-			String clientSecret = getAppSettings().getYoutubeClientSecret();
-
-			videoServiceEndpoint = getApplication().getEndpointService(AntMediaApplicationAdapter.YOUTUBE_ENDPOINT_CLASS, null, clientId, clientSecret);
-
-			if (isClientIdMissing(videoServiceEndpoint, clientId, clientSecret)) 
-			{
-				missingClientIdAndSecret = true;
-			}
-
-		}
-		else if (serviceName.equals(AntMediaApplicationAdapter.PERISCOPE)) 
-		{
-			String clientId = getAppSettings().getPeriscopeClientId();
-			String clientSecret = getAppSettings().getPeriscopeClientSecret();
-
-			videoServiceEndpoint = getApplication().getEndpointService(PeriscopeEndpoint.class.getName(), null, clientId, clientSecret);
-
-			if (isClientIdMissing(videoServiceEndpoint, clientId, clientSecret))  {
-				missingClientIdAndSecret = true;
-			}
-		}
-
-		try {
-
-			if (missingClientIdAndSecret) {
-				errorId = ERROR_SOCIAL_ENDPOINT_UNDEFINED_CLIENT_ID;
-				message = "Please enter service client id and client secret in app configuration";
-			}
-			else if (videoServiceEndpoint == null) {
-				errorId = ERROR_SOCIAL_ENDPOINT_UNDEFINED_ENDPOINT;
-				message = "Service with the name specified is not found in this app";
-			}
-			else {
-				DeviceAuthParameters askDeviceAuthParameters = videoServiceEndpoint.askDeviceAuthParameters();
-
-				getApplication().startDeviceAuthStatusPolling(videoServiceEndpoint,
-						askDeviceAuthParameters);
-				return askDeviceAuthParameters;
-			}
-		}
-		catch (Exception e) {
-			errorId = ERROR_SOCIAL_ENDPOINT_EXCEPTION_IN_ASKING_AUTHPARAMS;
-			message = "Exception in asking parameters";
-			logger.error(ExceptionUtils.getStackTrace(e));
-		}
-
-		return new Result(false, message, errorId);
-	}
-
-	private boolean isClientIdMissing(VideoServiceEndpoint videoServiceEndpoint, String clientId, String clientSecret) {
-		boolean result = false;
-		if ((videoServiceEndpoint != null) && 
-				(clientId == null || clientSecret == null || 
-				clientId.length() == 0 || clientSecret.length() == 0)) {
-			result = true;
-		}
-		return result;
+		return super.getDeviceAuthParameters(serviceName);
 	}
 
 	@ApiOperation(value = "Check if device is authenticated in the social network. In authorization phase, " +
@@ -618,40 +518,9 @@ public class BroadcastRestServiceV2 extends RestServiceBase{
 	@Consumes({ MediaType.APPLICATION_JSON })
 	@Path("/social-network-status/{userCode}")
 	@Produces(MediaType.APPLICATION_JSON)
+	@Override
 	public Result checkDeviceAuthStatus(@ApiParam(value = "Code of social media account", required = true) @PathParam("userCode") String userCode) {
-		Map<String, VideoServiceEndpoint> endPointMap = getEndpointList();
-		String message = null;
-		boolean authenticated = false;
-		String endpointId = null;
-		if (endPointMap != null) {
-			for (VideoServiceEndpoint videoServiceEndpoint : endPointMap.values()) {
-				//if there is an endpoint added to the list with same user code,
-				//it means it is authenticated
-				DeviceAuthParameters authParameters = videoServiceEndpoint.getAuthParameters();
-				if (authParameters != null && authParameters.user_code.equals(userCode)) {
-					authenticated = true;
-					endpointId = videoServiceEndpoint.getCredentials().getId();
-					break;
-				}
-			}
-		}
-		if (!authenticated) {
-			List<VideoServiceEndpoint> endPointList = getEndpointsHavingErrorList();
-			for (VideoServiceEndpoint videoServiceEndpoint : endPointList) {
-				DeviceAuthParameters authParameters = videoServiceEndpoint.getAuthParameters();
-				if (authParameters != null && authParameters.user_code.equals(userCode)) {
-					message = videoServiceEndpoint.getError();
-					endPointList.remove(videoServiceEndpoint);
-					break;
-				}
-			}
-
-		}
-		return new Result(authenticated, endpointId, message);
-	}
-
-	protected List<VideoServiceEndpoint> getEndpointsHavingErrorList(){
-		return getApplication().getVideoServiceEndpointsHavingError();
+		return super.checkDeviceAuthStatus(userCode);
 	}
 
 	@ApiOperation(value = "Get Credentials of Social Endpoints", notes = "", responseContainer = "List",response = SocialEndpointCredentials.class)
@@ -755,63 +624,6 @@ public class BroadcastRestServiceV2 extends RestServiceBase{
 		return result;
 	}
 
-	@Nullable
-	private Mp4Muxer getMp4Muxer(MuxAdaptor muxAdaptor) {
-		Mp4Muxer mp4Muxer = null;
-		for (Muxer muxer : muxAdaptor.getMuxerList()) {
-			if (muxer instanceof Mp4Muxer) {
-				mp4Muxer = (Mp4Muxer) muxer;
-			}
-		}
-		return mp4Muxer;
-	}
-
-	private Result startMp4Muxing(String streamId) {
-		boolean result = false;
-		List<MuxAdaptor> muxAdaptors = getMuxAdaptors(streamId);
-		for (MuxAdaptor muxAdaptor : muxAdaptors) {
-			if (muxAdaptor != null) {
-				Mp4Muxer mp4Muxer = getMp4Muxer(muxAdaptor);
-				if (mp4Muxer == null) {
-					//avoid multiple call of rest api adding new mp4muxers
-					muxAdaptor.startRecording();
-				}
-				result = true;
-			}
-		}
-		return new Result(result);
-	}
-
-	private List<MuxAdaptor> getMuxAdaptors(String streamId) {
-		AntMediaApplicationAdapter application = getApplication();
-		List<MuxAdaptor> muxAdaptors = new ArrayList<>();
-		if(application != null){
-			muxAdaptors = application.getMuxAdaptors();
-		}
-		List<MuxAdaptor> matchedMuxAdaptors = new ArrayList<>();
-		for (MuxAdaptor muxAdaptor : muxAdaptors) {
-			if (streamId.equals(muxAdaptor.getStreamId())) {
-				matchedMuxAdaptors.add(muxAdaptor);
-			}
-		}
-		return matchedMuxAdaptors;
-	}
-
-	private Result stopMp4Muxing(String streamId) {
-		boolean result = false;
-		List<MuxAdaptor> muxAdaptors = getMuxAdaptors(streamId);
-		for (MuxAdaptor muxAdaptor : muxAdaptors) {
-			if (muxAdaptor != null) {
-				Mp4Muxer mp4Muxer = getMp4Muxer(muxAdaptor);
-				if (mp4Muxer != null) {
-					//avoid multiple call of rest api stopping mp4 muxer
-					muxAdaptor.stopRecording();
-				}
-				result = true;
-			}
-		}
-		return new Result(result);
-	}
 
 	@ApiOperation(value = "Get IP Camera Error after connection failure", notes = "Notes here", response = Result.class)
 	@GET
